refactor: move react components to separate dir
