+++ src/datasource-zabbix/components/FunctionEditorControls.tsx
